ParseQri_Backend/ParseQri_Agent/TextToSQL_Agent/requirements.txt
pandas>=1.3.0
joblib>=1.0.0
plotly>=5.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
ollama>=0.1.0
numpy>=1.20.0
scikit-learn>=1.0.0
geopandas>=0.10.0
d3dshot>=0.1.0
chromadb>=0.4.18
sqlalchemy>=2.0.0
PyMySQL>=1.0.2
python-dotenv>=0.19.0
requests>=2.25.0 